new commit

+++ src/main/java/org/example/DoNewOrderOutput.java
@@ -8,9 +8,9 @@ public class DoNewOrderOutput {
     public int c_id;
     public String c_last;
     public String c_credit;
-    public int c_discount;
-    public int w_tax;
-    public int d_tax;
+    public Double c_discount;
+    public Double w_tax;
+    public Double d_tax;
     public int o_ol_cnt;
     public int o_id;
     public String o_entry_d;
